+++ src/views/inner-pages/dalrun-asj/inner/admin-bbs/AdminQuestion.jsx
@@ -1,11 +1,8 @@
 import AdminProductinquiry from "./admin-question/AdminProductinquiry";
 import AdminQna from "./admin-question/AdminQna";
 import AdminBtn from "../../../../../components/dalrun-asj/AdminBtn";
-import { Route, Routes, useParams } from "react-router-dom";
-import { useEffect, useState } from "react";
-import BasicSearch from "../../../../../components/dalrun-asj/BasicSearch";
+import { Route, Routes } from "react-router-dom";
 import AdminQuestionContent from "../admin-contents/AdminQuestionContent";
-import AdminSearch from "../../../../../components/dalrun-asj/AdminSerach";
 
 function AdminQuestion() {
     const qCategory = [
@@ -13,48 +10,15 @@ function AdminQuestion() {
         {cate: "qna", name:"QnA", selected:<AdminQna />}, 
     ];
 
-    const [optionVal, setOptionVal] = useState([]);
-    const { sub } = useParams();
-
-    const curr = () => {
-        if(sub === "productinquiry") {
-            const option = [
-                {value : "memId", name : "작성자"},
-                {value : "productId", name : "상품번호"},
-                {value : "title", name : "제목"},
-                {value : "content", name : "내용"},
-            ]
-
-            setOptionVal(option);
-        } else {
-            const option = [
-                {value : "memId", name : "작성자"},
-                {value : "category", name : "카테고리"},
-                {value : "question", name : "질문"},
-            ]
-
-            setOptionVal(option);
-        }
-    }
-    
-    
-    useEffect(()=>{
-        curr();
-    }, [sub]);
-    
+   
     return (
         <>  
             <div style={{ float:"right" }}>
                 <AdminBtn {...qCategory}/>
             </div>
-            <div className="bbs">
-                <div className="bbs-content">
-                    <BasicSearch {...optionVal} />
-                    <Routes>
-                        <Route path=":sub/*" element={<AdminQuestionContent {...qCategory} />} />
-                    </Routes>
-                </div>
-            </div>
+            <Routes>
+                <Route path=":sub/*" element={<AdminQuestionContent {...qCategory} />} />
+            </Routes>
         </>
     );
 }
